+++ 0x03-debugging/main.h
@@ -1,6 +1,6 @@
 #ifndef MAIN_H
 #define MAIN_H
 #include <stdio.h>
-int largest_number(int, int, int);
+double largest_number(double, double, double);
 void positive_or_negative(int i);
 #endif
